Refactor Paper.estimate_pubdate
Moving this method into lib/arxiv
