+++ .gitignore
@@ -16,6 +16,7 @@
 /log/*.log
 /log/*.pid
 /tmp
+/coverage
 
 # Ignore local config files
 config/database.yml
